refactor(media): Refactored Media and RichText Media block to keep code DRY
